Refactor NavBar
