+++ src/components/SellerAccount/SellerAccount.js
@@ -2,7 +2,7 @@ import React, { useState, useEffect } from 'react'
 import { Card, Button } from 'antd'
 import axios from 'axios'
 import './SellerAccount.css'
-import Nav from '../inventory/nav'
+import Nav from '../elements/nav'
 import withAuth from '../Auth/axiosWithAuth'
 
 const storeURL = 'https://shopping-cart-eu3.herokuapp.com/api/store'
